requirements.txt
langchain
langchain-community
langchain_openai
numexpr
python-dotenv
streamlit
watchdog
wikipedia
youtube-transcript-api
faiss-cpu